Clean and organize submit-photo-upload-form and submit-z-score-form e2e test
