+++ tests/e2e/default/enketo/forms/z-score.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0"?>
+<h:html xmlns="http://www.w3.org/2002/xforms" xmlns:h="http://www.w3.org/1999/xhtml" xmlns:ev="http://www.w3.org/2001/xml-events" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:jr="http://openrosa.org/javarosa">
+    <h:head>
+        <h:title>z-score() - Demo Form</h:title>
+
+        <model>
+            <instance>
+                <data id="z-score">
+                    <my_height/>
+                    <my_weight/>
+                    <my_sex/>
+                    <my_age/>
+                    <wfh/>
+                    <wfa/>
+                    <hfa/>
+                    <meta>
+                        <instanceID/>
+                    </meta>
+                </data>
+            </instance>
+
+            <bind nodeset="/data/my_height"/>
+            <bind nodeset="/data/my_weight"/>
+            <bind nodeset="/data/my_age"/>
+            <bind nodeset="/data/my_sex"/>
+            <bind nodeset="/data/hfa" type="string" calculate="z-score('height-for-age', ../my_sex, ../my_age, ../my_height)" readonly="true()"/>
+            <bind nodeset="/data/wfa" type="string" calculate="z-score('weight-for-age', ../my_sex, ../my_age, ../my_weight)" readonly="true()"/>
+            <bind nodeset="/data/wfh" type="string" calculate="z-score('weight-for-height', ../my_sex, ../my_height, ../my_weight)" readonly="true()"/>
+        </model>
+    </h:head>
+
+    <h:body>
+        <group appearance="field-list" ref="/data">
+            <select1 appearance="horizontal" ref="/data/my_sex">
+                <label>Gender</label>
+                <item>
+                    <label>Female</label>
+                    <value>female</value>
+                </item>
+                <item>
+                    <label>Male</label>
+                    <value>male</value>
+                </item>
+            </select1>
+            <input ref="/data/my_height">
+                <label>How tall are you? (cm)</label>
+            </input>
+            <input ref="/data/my_weight">
+                <label>How much do you weigh? (kg)</label>
+            </input>
+            <input ref="/data/my_age">
+                <label>How old are you? (days)</label>
+            </input>
+            <input ref="/data/hfa">
+                <label>height for age</label>
+            </input>
+            <input ref="/data/wfa">
+                <label>weight for age</label>
+            </input>
+            <input ref="/data/wfh">
+                <label>weight for height</label>
+            </input>
+        </group>
+    </h:body>
+</h:html>
